Make `Verified` fields private
